Atualizações realizadas no projeto site de animes

+++ src/main/java/com/animeplus/dtos/AvaliacaoDTO.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AvaliacaoDTO {
-    private Integer id;
-    private Integer animeId;
-    private Integer episodioId;
+    private Long id;
+    private Long animeId;
+    private Long episodioId;
     private Double notaUsuario;
     private String comentario;
 }
